Refactor CLI code: Extract utility functions into separate modules
